extracts kpi scoring fns into module
