+++ epilepsy12/common_view_functions/calculate_kpi_functions/check_is_registered.py
@@ -1,0 +1,15 @@
+# python imports
+
+# django imports
+
+# E12 imports
+
+
+def check_is_registered(registration_instance) -> bool:
+    """
+    Helper fn returns True if registered
+    """
+    return (
+        registration_instance.registration_date is not None
+        and registration_instance.eligibility_criteria_met
+    )
